requirements.txt
spidev==3.5
RPi.GPIO==0.7.0
pillow==8.1.0